Add files via upload
This is the work I did in week 1 of learning elm with the 3d scenes and physics library.
IntersectingQuad has two Quadrilaterals rotated in along z axis and intersected.
CuboidTry was a basic attempt at rendering a static sphere and a static cuboid
TryCircle has a sphere that responds to mouse movements
BallBuil has a ball and a building, the ball responds to mouse events and you can throw the ball at the building
Wicket has 3 cricket stumps and a red colored cricket ball. You can throw the ball at the wickets and they'll fall
ThreeWalls was my attempt at changing the camera angle and focusing on the bigger picture
SeeSaw was my highlight of the week. Evident from the name it has a seesaw-like structure with a block on one end. On the empty end you can throw a ball so that the block bounces and gets thrown depending on how hard you throw the ball