Add files via upload
The second week was very challenging
I had to learn how to move 3D objects with keyboard arrow keys.
I was able to talk to the creators of the 3D scenes package and Physics package too. They were very helpful and wanted to solve all the issues.
KeyBall was an attempt at moving the ball with arrow keys, I feel the code is almost done, just some little errors
Mae is a maze that I created by understanding the weird coordinate system, putting the camera on top, and moving all blocks under one body and making it static. In this, you can move the ball with mouse and clear the maze.
MaeFol was supposed to be the modification of the maze in which the camera would follow the Ball. I can really use some help here. It's almost done.